cv/requirements.txt
opencv-python==4.3.0.36
tensorflow-datasets==3.2.1
pytest==6.0.1
